requirements.txt
Babel==2.12.1
Flask==2.3.2
Flask_Bootstrap==3.3.7.1
matplotlib==3.7.2
numpy==1.22.4
ossapi==3.3.1
pandas==2.0.3
python_dateutil==2.8.2
rosu_pp_py==0.9.4
schedule==1.2.0
scipy==1.11.1
seaborn==0.12.2
